Split clocks into (more) modules
